requirements.txt
# Runtime dependencies / Runtime-afhankelijkheden
mysql-connector-python>=8.3.0
pymodbus>=3.6.0
requests>=2.31.0
anthropic>=0.25.0
APScheduler>=3.10.0
Flask>=3.0.0
